Consolidate contant files and add index for SearchDialect
