--- frontend/app/state/reducers/exportDialect/constants.js
@@ -1,10 +1,0 @@
-// https://nuxeo.github.io/api-playground/#/commands/Document.GetExportProgress
-// https://nuxeo.github.io/api-playground/#/commands/Document.GetFormattedDocument
-// https://nuxeo.github.io/api-playground/#/commands/Document.FVGenerateDocumentWithFormat
-
-export const EXPORT_DEFAULT = 'EXPORT_DEFAULT'
-export const EXPORT_ERROR = 'EXPORT_ERROR'
-export const EXPORT_IN_PROGRESS = 'EXPORT_IN_PROGRESS'
-export const EXPORT_INITIALIZING = 'EXPORT_INITIALIZING'
-export const EXPORT_SUCCESS = 'EXPORT_SUCCESS'
-export const EXPORT_BROWSE = 'EXPORT_BROWSE'
