+++ frontend/app/state/reducers/searchDialect/reducer.js
@@ -13,7 +13,7 @@ import {
   SEARCH_TYPE_STARTS_WITH_SEARCH,
   SEARCH_TYPE_ENDS_WITH_SEARCH,
   SEARCH_TYPE_WILDCARD_SEARCH,
-} from 'components/SearchDialect/constants'
+} from 'common/Constants'
 
 export const initialState = {
   searchByAlphabet: '',
